final_project/task_manager/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
requests = "^2.26.0"
uvicorn = "^0.15.0"
retry = "^0.9.2"
loguru = "^0.5.3"
phonenumbers = "^8.12.30"
pydantic = {version = "^1.8.2", extras = ["email"]}
dependency-injector = "^4.35.2"
SQLAlchemy = "1.4.23"
asgi-lifespan = "^1.0.1"
SQLAlchemy-Utils = "^0.37.8"
pre-commit = "^2.14.0"
httpx = "^0.19.0"
aiomysql = "^0.0.21"
sentry-sdk = "^1.3.1"
uvloop = "^0.16.0"
starlette-prometheus = "^0.7.0"
python-dotenv = "^0.19.0"
flakehell = "0.9.0"
factory-boy = "^3.2.0"
sqlakeyset = "^1.0.1629029818"
python-jose = "^3.3.0"
aiokafka = "^0.7.2"
asyncpg = "^0.24.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.15.1"
bumpversion = "^0.6.0"
isort = "^5.9.3"
black = "^21.7b0"
flake8-logging-format = "^0.6.0"
flake8-comprehensions = "^3.6.1"
flake8-eradicate = "^1.1.0"
flake8-pytest-style = "^1.5.0"
flakehell = "^0.9.0"
flake8 = "^3.9.2"
flake8-import-order = "^0.18.1"
flake8-aaa = "^0.12.0"
flake8-bugbear = "^21.4.3"
flake8-debugger = "^4.0.0"
flake8-expression-complexity = "^0.0.9"
flake8-fixme = "^1.1.1"
flake8-simplify = "^0.14.1"
flake8-variables-names = "^0.0.4"
flake8-bandit = "^2.1.2"
pytest-deadfixtures = "^2.2.1"
pytest-testmon = "^1.1.2"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.1"
pytest-timeout = "^1.4.2"
hypothesis = "^6.15.0"
ipython = "^7.26.0"
safety = "^1.10.3"
yamllint = "^1.26.3"
tomlkit = "^0.7.2"
pytest-sugar = "^0.9.4"
mypy = "^0.910"
pre-commit = "^2.14.0"
pytest = "^6.2.4"
flake8-tidy-imports = "^4.4.1"
pyupgrade = "^2.24.0"




[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38', 'py39']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docker
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[tool.isort]
profile = "black"
skip = "setup.py,env,node_modules,.tox, venv, .venv,"

[tool.flakehell]
extended_default_ignore = []
exclude = ["README.md", ".venv", ]
format = "grouped"
max_line_length = 120
show_source = true

[tool.flakehell.plugins]
pyflakes = ["+*", "-F401"]
flake8-aaa = ["-*"]
flake8-bandit = ["-*", "+S1??"]
flake8-bugbear = ["+*", "-B008"]
flake8-comprehensions = ["+*"]
flake8-debugger = ["+*"]
flake8-expression-complexity = ["-*"]
flake8-fixme = ["-*"]
flake8-eradicate = ["+*"]
flake8-simplify = ["+*"]
flake8-pytest-style = ["+*"]
flake8-tidy-imports = ["+*"]


[tool.flakehell.exceptions."tests/"]
flake8-bandit = ["-*"]
flake8-comprehensions = ["+*"]
flake8-eradicate = ["+*"]
flake8-pytest-style = ["+*"]
mccabe = ["-*"]
pep8-naming = ["-*"]
pyflakes = ["-*"]


[tool.flakehell.exceptions."gunicorn_config.py"]
flake8-bandit = ["-*"]
flake8-comprehensions = ["+*"]
flake8-eradicate = ["+*"]
flake8-pytest-style = ["+*"]
mccabe = ["-*"]
pep8-naming = ["-*"]
pyflakes = ["-*"]


[tool.pytest.ini_options]
minversion = "6.0"
testpaths = "tests"
filterwarnings = '''
  error
  ignore::DeprecationWarning
'''
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=90
'''
